Some more GUI stuff, mostly coded for testing things right now
CompositeTimeline only receives an arraylist of tweets and populates enough tweetcontainers for the compositetimeline jframe
SubscriptionList is blank but it will be a list of our subscribers on the left hand side of the main twitterapp
TweetContainer takes in a tweet object (from compositetimeline) and grabs all the information from whatever tweet was passed to it.

TwitterApp now creates a new JFrame and populates a contentpane with a composite timeline for viewing.  Works.
Tweeter was just changed to parse XML for a userpicture, but haven't been able to add the userpicture of type Image to a TweetContainer GUI object, yet.

That's All, Folks.